split device into separate modules
