+++ release-checklist.md
@@ -1,0 +1,8 @@
+# Pre-release checklist
+
+Make sure to do the following, before publishing a new release:
+
+- Search for "TODO/rel:" comments. These are to be done before publishing next release.
+- Run all tests.
+- Run all examples.
+- Validate the changelog, make sure it's populated with the latest changes.
